pyproject.toml
[project]
name = "conductor-fastapi-distributed"
version = "0.1.0"
description = ""
authors = [
    {name = "Tushar Chopra",email = "tusharchopratech@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "conductor-python (>=1.1.10,<2.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
